Deleted imports and changed 'windDirection' in PortraitWeatherUI composable from Int to Float

+++ code/CoolJetpackWeatherApp/app/src/main/java/dam_A51696/cooljetpackweatherapp/ui/WeatherScreen.kt
@@ -15,14 +15,10 @@ import dam_A51696.cooljetpackweatherapp.viewmodel.WeatherViewModel
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.*
-import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
@@ -106,7 +102,7 @@ fun PortraitWeatherUI(
     longitude: Float,
     temperature: Float,
     windSpeed: Float,
-    windDirection: Int,
+    windDirection: Float,
     weathercode: Int,
     seaLevelPressure: Float,
     time: String,
@@ -163,7 +159,7 @@ fun LandscapeWeatherUI(
     longitude: Float,
     temperature: Float,
     windSpeed: Float,
-    windDirection: Int,
+    windDirection: Float,
     weathercode: Int,
     seaLevelPressure: Float,
     time: String,
